Refactor: Organize remaining views into subdirectories
