+++ src/views/Auth/Login.vue
@@ -20,7 +20,7 @@
 
 <script setup>
 import { ref } from "vue";
-import { useAuthStore } from "../stores/auth.js";
+import { useAuthStore } from "../../stores/auth.js";
 import { message } from "ant-design-vue";
 import { useRouter } from "vue-router";
