+++ src/views/Organization/Organizations.vue
@@ -1,7 +1,7 @@
 <script setup>
 import { ref, onMounted, computed } from 'vue';
 import { useRouter } from 'vue-router';
-import { useOrganizationsStore } from '../stores/organizations';
+import { useOrganizationsStore } from '../../stores/organizations';
 import { message } from 'ant-design-vue';
 import { PlusOutlined } from '@ant-design/icons-vue';
