+++ src/views/Staff/Users.vue
@@ -42,8 +42,8 @@ import { ref, onMounted } from 'vue';
 import axios from 'axios';
 import { message } from 'ant-design-vue';
 import { PlusOutlined } from '@ant-design/icons-vue';
-import { useAuthStore } from '../stores/auth';
-import AddUserDrawer from '../components/AddUserDrawer.vue';
+import { useAuthStore } from '../../stores/auth';
+import AddUserDrawer from '../../components/AddUserDrawer.vue';
 
 const users = ref([]);
 const loading = ref(false);
